removed scroll bar, removed rerender of chapter menu
