+++ src/components/css/SideMenu.module.css
@@ -64,6 +64,7 @@ body.sideMenuOpen {
 
 .itemsList::-webkit-scrollbar {
     height: 0;
+    width: 0;
 }
 
 .button {
